requirements.txt
Django==2.0.7
django-crontab==0.7.1
django-filter==1.1.0
django-guardian==1.4.9
django-redis==4.8.0
django-reversion==2.0.13
djangorestframework==3.7.7
djangorestframework-bulk==0.2.1
djangorestframework-filters==0.10.2
djangorestframework-jwt==1.11.0
mysqlclient==1.3.10
Pillow==5.2.0
pycparser==2.18
PyJWT==1.6.4
redis==2.10.6
requests==2.19.1
rest-condition==1.0.3
gunicorn
gevent